move most webconsole plugins to console (ClientlibDebugConsolePlugin stays where it is since I want all clientlib stuff together.)
